+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Fixed
+- Enforce CDC chunk size ordering in handshake validation.
 - validate block size mismatch in handshakes
 
 ## [v0.1.0] - 2025-02-27